divide class/struct
