+++ JuiceMaker/JuiceMaker/Fruit.swift
@@ -1,0 +1,38 @@
+//
+//  Fruit.swift
+//  JuiceMaker
+//
+//  Created by Wonhee on 2020/11/19.
+//
+
+import Foundation
+
+enum FruitsType : String {
+    case strawberry = "🍓"
+    case banana = "🍌"
+    case pineapple = "🍍"
+    case kiwi = "🥝"
+    case mango = "🥭"
+}
+
+class Fruit {
+    let fruitType: FruitsType
+    private(set) var stock: Int
+    
+    init(fruitType: FruitsType) {
+        self.fruitType = fruitType
+        self.stock = 10
+    }
+    
+    func addStock(_ add: Int) {
+        self.stock = self.stock + add
+    }
+    
+    func useStock(_ use: Int) {
+        self.stock = self.stock - use
+    }
+    
+    func canMakeJuice(need: Int) -> Bool {
+        return stock >= need
+    }
+}
